Added public APIs for Malloc and Free. Free is needed to free memory created by API functions like Print().
The problem is that when multiple libraries/users use CJSON in a larger
application context, it is not
always known if a custom allocator was supplied. Thus a way to call the
custom free function is required.

Malloc is made public also for completeness/balance, however, I don't think
there is ever a case that it is needed.